Remove the double-close in node, gateway, and registration (permissioning).
The Serve(...) function documentation specifies that it listens until the port is closed, and the GracefulStop() function is supposed to close the port. We were seeing errors in logs with the ports being closed twice, and we think it could be releated to the connection issues we see in the integration test. It is not clear this will fix the error we are seeing, but the additional log message will make it clearer when the server is stopping and the port should be closed, and this change will bring us into sync with the current golang documentation on how the functions work, specifically:

' Serve accepts incoming connections on the listener lis, creating a new ServerTransport and service goroutine for each. The service goroutines read gRPC requests and then call the registered handlers to reply to them. Serve returns when lis.Accept fails with fatal errors. lis will be closed when this method returns. Serve will return a non-nil error unless Stop or GracefulStop is called.
'