Update RegisterNode comm to use ports, not addresses
The RegisterNode comm initially used GatewayAddress, which includes a hostname:port, and the node:port
component was inferred by the handler via the connection. This did not work because the port derived
is either an ephemeral port or the listener of the registration server.

The effect is, in production, with everything using the same ports, it could work, but you could not
test locally, as the ports needed to be different to run multiple nodes and a registration server
simultaneously.