Update dependencies.
Note that protobufs and related are held back because theres so much deprecated code that would be affected. There are also now a number of locking and other issues (e.g., we've still got sidh in this branch) that are due for cleanup but for architecture support or legacy support reasons we cannot touch at this time.